requirements.txt
streamlit==1.21.0
langchain
langchain-community
langchain-openai
langchain-groq
mysql-connector-python
python-dotenv
altair==4.2.0